// lib/src/types.ts
export interface BaseData {
  id?: string | null;
  _id?: string | null;
  [key: string]: any;
}

export type PathType<RowType extends BaseData> = Exclude<LiteralUnion<keyof RowType, string>, symbol | number>;

export type PathValueType<T extends BaseData> = PathType<T> | true | undefined;

export type Sorter<T> = (ab: T, ba: T) => number;

export type TableCellAlign = "left" | "center" | "right";

export type DataTypes = typeof DATA_TYPES[number];

export type NullableDataTypes = DataTypes | undefined | null;
// ...
export type OperatorValues = typeof BASE_OPERATORS[number]["value"];
// ...
export type FilterValue = string | number | Date | boolean | null;

export interface ActiveFilter<RowType extends BaseData = any> {
  id: string;
  type: NonNullable<NullableDataTypes>;
  path: PathType<RowType>;
  value: FilterValue;
  operator: OperatorValues;
}

export type NullableActiveFilter<RowType extends BaseData = any> = {
  [key in keyof ActiveFilter<RowType>]: ActiveFilter<RowType>[key] | null;
};